+++ src/table/Table.js
@@ -298,8 +298,7 @@ export const Table = ({
   fromIndex,
   customProps
 }) => {
-  const filterable = R.pathOr(true, [modelName, 'filterable'], schema)
-  const allColFilterable = isTableFilterable({schema, modelName, tableView})
+  const allColFilterable = isTableFilterable({schema, modelName, user})
 
   if (!fromIndex && hideTable) {
     return null
@@ -311,7 +310,6 @@ export const Table = ({
   const deletable = isTableDeletable({ schema, modelName, data, parentNode, user, customProps })
   const detailField = calcDetailField({schema, modelName, fieldOrder})
   const editable = isTableEditable({ schema, modelName, data, parentNode, user, fieldOrder, customProps })
-  const sortable = R.pathOr(true, [modelName, 'sortable'], schema)
 
   return (
     <table className='table table-striped table-bordered table-hover'>
@@ -324,11 +322,10 @@ export const Table = ({
         editable,
         detailField,
         selectOptions,
-        sortable,
-        filterable,
         tableView,
         fromIndex,
-        customProps
+        customProps,
+        user
       }} />
       <Body {...{
         schema,